capital title
I noticed before my submission that the resume and portfolio pages were breaking when I tried to open them on my phone. I spent hours trying to figure out how to fix it, but I was leaving for a convention in South Carolina on Thursday night so after 5p.m. I had to give up. I had check my "sites" folder to make sure the pages were correct etc. and just could not figure it out. Friday, while I was in South Carolina, I finally figured out that the pages were capitalized on github.com and tried to fix it but my friends computer wouldn't let me for some reason. So, now I am home and on my computer and its letting me so Im gonna go ahead and change the capitalized pages. I know we are not supposed to make any changes past the due date, so I understand if this is not considered in my grade, but I just wanted to go ahead and change it so you can at least see the page.